requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug>=1.0.1
bcrypt==3.1.7
azure-cosmos == 4.0.0
marshmallow == 3.6.1
Flask-Login == 0.5.0
flask-marshmallow == 0.13.0
uuid == 1.30
Flask-Mail == 0.9.1
flask-wtf == 0.14.3
dnspython == 1.16.0
email-validator == 1.1.1
Flask-Bcrypt == 0.7.1
requests == 2.24.0
DateTime == 4.3
finnhub-python == 2.2.0
pandas == 1.1.1
mplfinance == 0.12.7a0
matplotlib == 3.3.1
pybase64 == 1.0.1
loxun == 2.0